생명구현(?)
바뀐숫자가 안사라짐(오류)

+++ Framework/Framework/ImageResize.h
@@ -4,6 +4,6 @@ class ImageResize
 {
 public:
 	double width, height;
-	void resize(GameObject* g, double x, double y);
+	void resize(GameObject* g, int x, int y);
 };
